Replace sha1 with sha256
Existing fingerprints in authorized_certs file in the beginning of every
line will not be updated, that fingerprint is not used for anything.

It's just used to help the user know what the fingerprint is.

I could not think of a great way to update the fingerprints already
there without adding alot of code, which I think would be a waste.

To get the fingerprint updated the user would need to delete that
line or the whole file and exchange/pair again.