Set notification desktop-entry hint
Apparently KDE does not keep notifications in history if the
notification is not configurable (the desktop-entry hint must be set).

Quote from
https://cgit.kde.org/plasma-workspace.git/commit/?id=1f6050b1740cf800cb031e98fd89bc00ca20c55a

"If a notification isn't configurable (e.g. because the application
doesn't set a desktop-entry hint), there's no way for the user to keep
it out of the history if they so desire, causing excess history spam."

Closes #46